+++ script.js
@@ -1,16 +1,18 @@
-import { fetchGet } from "./api.js";
-const container = document.querySelector('.container');
-const form = document.querySelector('.add-form');
-const inputName = form.querySelector('.add-form-name');
-const inputComment = form.querySelector('.add-form-text');
-const formButton = form.querySelector('.add-form-button');
-const listComments = document.querySelector('.comments');
-const error = document.querySelector('.error');
-const loading = document.querySelector('.load');
-let inputNameValue = '';
-let inputCommentValue = '';
+import { fetchGet, fetchPost } from "./api.js";
+import { inputName, inputComment, listComments, formButton, form, inputNameValue, inputCommentValue } from "./variables.js";
+import { renderDate } from './render.js';
+import { renderLoading, removerLoading } from "./handlerLoading.js" ;
 
-let comments = [];
+const container = document.querySelector('.container');
+// const form = document.querySelector('.add-form');
+// const inputName = form.querySelector('.add-form-name');
+// const inputComment = form.querySelector('.add-form-text');
+// const formButton = form.querySelector('.add-form-button');
+// const listComments = document.querySelector('.comments');
+const error = document.querySelector('.error');
+// const loading = document.querySelector('.load');
+// let inputNameValue = '';
+// let inputCommentValue = '';
 
 
 // // Функция запроса GET
@@ -54,58 +56,58 @@ let comments = [];
 
 
 
-// Функция запроса POST
-const fetchPost = (newComment) => {
-	fetch('https://webdev-hw-api.vercel.app/api/v1/julia-matsievich/comments', {
-		method: 'POST',
-		body: JSON.stringify(newComment)
-	})
-		.then(response => {
-			if (response.status === 400) {
-				throw new Error('код 400');
-			}
-			if (response.status === 500) {
-				throw new Error('код 500');
-			}
-			return response.json();
-		})
-		.then(responseData => {
-			comments = responseData.comments;
-			fetchGet()
-		})
-		.catch(error => {
-			removerLoading();
-			inputName.value = inputNameValue;
-			inputComment.value = inputCommentValue;
-			if (error.message === 'код 400') {
-				alert('Имя и комментарий должны быть не менее 3х символов');
-				return;
-			}
-			if (error.message === 'код 500') {
-				fetchPost(newComment);
-				initialState();
-			} else {
-				alert('Кажется, у вас сломался интернет, попробуйте позже')
-				console.log(error);
-			}
-		})
-}
+// // Функция запроса POST
+// const fetchPost = (newComment) => {
+// 	fetch('https://webdev-hw-api.vercel.app/api/v1/julia-matsievich/comments', {
+// 		method: 'POST',
+// 		body: JSON.stringify(newComment)
+// 	})
+// 		.then(response => {
+// 			if (response.status === 400) {
+// 				throw new Error('код 400');
+// 			}
+// 			if (response.status === 500) {
+// 				throw new Error('код 500');
+// 			}
+// 			return response.json();
+// 		})
+// 		.then(responseData => {
+// 			comments = responseData.comments;
+// 			fetchGet()
+// 		})
+// 		.catch(error => {
+// 			removerLoading();
+// 			inputName.value = inputNameValue;
+// 			inputComment.value = inputCommentValue;
+// 			if (error.message === 'код 400') {
+// 				alert('Имя и комментарий должны быть не менее 3х символов');
+// 				return;
+// 			}
+// 			if (error.message === 'код 500') {
+// 				fetchPost(newComment);
+// 				initialState();
+// 			} else {
+// 				alert('Кажется, у вас сломался интернет, попробуйте позже')
+// 				console.log(error);
+// 			}
+// 		})
+// }
 
 
 
 
 
-//Создание даты в нужном формате
-function renderDate(dataDate) {
-	const date = new Date(dataDate);
-	const dateDataArr = date.toLocaleDateString().split('.')
-	dateDataArr[dateDataArr.length - 1] = dateDataArr[dateDataArr.length - 1].slice(2);
-	const dateData = dateDataArr.join('.');
+// //Создание даты в нужном формате
+// function renderDate(dataDate) {
+// 	const date = new Date(dataDate);
+// 	const dateDataArr = date.toLocaleDateString().split('.')
+// 	dateDataArr[dateDataArr.length - 1] = dateDataArr[dateDataArr.length - 1].slice(2);
+// 	const dateData = dateDataArr.join('.');
 
-	const time = date.toLocaleTimeString().slice(0, -3);
+// 	const time = date.toLocaleTimeString().slice(0, -3);
 
-	return `${dateData} ${time}`;
-}
+// 	return `${dateData} ${time}`;
+// }
 
 //Показать ошибку
 function showError(element) {
@@ -133,7 +135,7 @@ function isValid() {
 }
 
 //Начальное состояние формы
-function initialState() {
+export function initialState() {
 	inputName.value = '';
 	inputComment.value = '';
 	formButton.disabled = true;
@@ -142,8 +144,8 @@ function initialState() {
 
 // Обработка добавления комментария
 function handlerAddComment() {
-	inputNameValue = inputName.value;
-	inputCommentValue = inputComment.value;
+	let inputNameValue = inputName.value;
+	let inputCommentValue = inputComment.value;
 
 	hideError();
 	if (!isValid()) {
@@ -171,199 +173,199 @@ function handlerAddComment() {
 	initialState();
 }
 
-//Создание формы редактирования комментария
-function renderEditComment(element) {
-	const parent = element.closest('.comment');
-	const commentBody = parent.querySelector('.comment-body');
-	const commentText = parent.querySelector('.comment-text');
-	const divQuote = commentText.querySelector('.quote');
-	const editTextarea = commentText.innerHTML.replaceAll('<div class="quote">', '/**').replaceAll('</div>', '**/');
-	commentBody.innerHTML = `
-	<div class="edit-form">
-	<textarea type="textarea" class="edit-form-text" rows="4">${editTextarea}</textarea>
-	<div class="add-form-row">
-	  <button class="edit-form-button">Сохранить</button>
-	</div>
-	</div>
-	`
-	const editForm = parent.querySelector('.edit-form');
+// //Создание формы редактирования комментария
+// function renderEditComment(element) {
+// 	const parent = element.closest('.comment');
+// 	const commentBody = parent.querySelector('.comment-body');
+// 	const commentText = parent.querySelector('.comment-text');
+// 	const divQuote = commentText.querySelector('.quote');
+// 	const editTextarea = commentText.innerHTML.replaceAll('<div class="quote">', '/**').replaceAll('</div>', '**/');
+// 	commentBody.innerHTML = `
+// 	<div class="edit-form">
+// 	<textarea type="textarea" class="edit-form-text" rows="4">${editTextarea}</textarea>
+// 	<div class="add-form-row">
+// 	  <button class="edit-form-button">Сохранить</button>
+// 	</div>
+// 	</div>
+// 	`
+// 	const editForm = parent.querySelector('.edit-form');
 
-	editForm.addEventListener('click', (event) => {
-		event.stopImmediatePropagation()
-	})
-	const newCommentText = parent.querySelector('.edit-form-text');
-	const editButtonSave = parent.querySelector('.edit-form-button');
-	const commentIndex = parent.dataset.comment;
+// 	editForm.addEventListener('click', (event) => {
+// 		event.stopImmediatePropagation()
+// 	})
+// 	const newCommentText = parent.querySelector('.edit-form-text');
+// 	const editButtonSave = parent.querySelector('.edit-form-button');
+// 	const commentIndex = parent.dataset.comment;
 
-	editButtonSave.addEventListener('click', (event) => {
-		event.stopPropagation();
-		if (!newCommentText.value) {
-			showError(commentBody);
-			return;
-		}
+// 	editButtonSave.addEventListener('click', (event) => {
+// 		event.stopPropagation();
+// 		if (!newCommentText.value) {
+// 			showError(commentBody);
+// 			return;
+// 		}
 
-		const newCommentTextValue = newCommentText.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("/**", "<div class='quote'>").replaceAll("**/", "</div>");
+// 		const newCommentTextValue = newCommentText.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("/**", "<div class='quote'>").replaceAll("**/", "</div>");
 
-		if (divQuote) {
-			commentBody.innerHTML = `<div class="comment-text">${newCommentTextValue}</div>`
-		} else {
-			commentBody.innerHTML = `<div class="comment-text">${newCommentTextValue}</div>`
-		}
+// 		if (divQuote) {
+// 			commentBody.innerHTML = `<div class="comment-text">${newCommentTextValue}</div>`
+// 		} else {
+// 			commentBody.innerHTML = `<div class="comment-text">${newCommentTextValue}</div>`
+// 		}
 
-		comments[commentIndex].text = newCommentTextValue;
-	})
-}
+// 		comments[commentIndex].text = newCommentTextValue;
+// 	})
+// }
 
-//Рендер комментариев
-function renderComments() {
+// //Рендер комментариев
+// function renderComments() {
 
-	const commentsHtml = comments.map((comment, index) => {
-		comment.likeButtonClass = '';
+// 	const commentsHtml = comments.map((comment, index) => {
+// 		comment.likeButtonClass = '';
 
-		if (comment.isLiked) {
-			comment.likeButtonClass = '-active-like';
-		}
+// 		if (comment.isLiked) {
+// 			comment.likeButtonClass = '-active-like';
+// 		}
 
-		return `
-			<li class="comment" data-comment=${index}>
-			<div class="comment-header">
-			  <div>${comment.name}</div>
-			  <div>${comment.date}</div>
-			</div>
-			<div class="comment-body">
-			  <div class="comment-text">${comment.text}</div>
-			</div>
-			<div class="comment-footer">
-			  <div class="likes">
-				 <span class="likes-counter" data-likeCounter = "${comment.likes}">${comment.likes}</span>
-				 <button class="like-button ${comment.likeButtonClass}"></button>
-				 <button class="edit-button"></button>
-				 <button class="delete-button" data-index="${index}"></button>
-			  </div>
-			</div>
-		 </li>`
+// 		return `
+// 			<li class="comment" data-comment=${index}>
+// 			<div class="comment-header">
+// 			  <div>${comment.name}</div>
+// 			  <div>${comment.date}</div>
+// 			</div>
+// 			<div class="comment-body">
+// 			  <div class="comment-text">${comment.text}</div>
+// 			</div>
+// 			<div class="comment-footer">
+// 			  <div class="likes">
+// 				 <span class="likes-counter" data-likeCounter = "${comment.likes}">${comment.likes}</span>
+// 				 <button class="like-button ${comment.likeButtonClass}"></button>
+// 				 <button class="edit-button"></button>
+// 				 <button class="delete-button" data-index="${index}"></button>
+// 			  </div>
+// 			</div>
+// 		 </li>`
 
-	})
+// 	})
 
-	listComments.innerHTML = commentsHtml.join('');
-	initLikeButtonEventListeners();
-	initEditButtonEventListeners();
-	initDeleteButtonEventListeners();
-	initAnswerCommentEventListener();
-}
+// 	listComments.innerHTML = commentsHtml.join('');
+// 	initLikeButtonEventListeners();
+// 	initEditButtonEventListeners();
+// 	initDeleteButtonEventListeners();
+// 	initAnswerCommentEventListener();
+// }
 
-//Рендер загрузки
-function renderLoading() {
-	form.classList.add('hidden');
-	loading.classList.remove('hidden');
-}
+// //Рендер загрузки
+// function renderLoading() {
+// 	form.classList.add('hidden');
+// 	loading.classList.remove('hidden');
+// }
 
-//Убрать загрузку
-function removerLoading() {
-	form.classList.remove('hidden');
-	loading.classList.add('hidden');
-}
-
-
-// Подписка на события кнопки Редактировать
-function initEditButtonEventListeners() {
-	const EditButtons = listComments.querySelectorAll('.edit-button');
-
-	for (const EditButton of EditButtons) {
-		EditButton.addEventListener('click', (event) => {
-			event.stopPropagation();
-			const target = event.target;
-			renderEditComment(target)
-		})
-	}
-}
-
-//Подписка на события клика по кнопке Лайк
-function initLikeButtonEventListeners() {
-
-	const likeButtons = listComments.querySelectorAll('.like-button');
-
-	for (const likeButton of likeButtons) {
-		likeButton.addEventListener('click', (event) => {
-			event.stopPropagation();
-			const target = event.target;
-			const commentParrent = target.closest('.comment');
-			const commentId = commentParrent.dataset.comment;
-			likeButton.classList.add('-loading-like');
-
-			delay(2000).then(() => {
-				comments[commentId].likes = comments[commentId].isLiked ? comments[commentId].likes -= 1 : comments[commentId].likes += 1;
-				comments[commentId].isLiked = !comments[commentId].isLiked;
-				renderComments()
-			});
-		})
-	}
-}
+// //Убрать загрузку
+// function removerLoading() {
+// 	form.classList.remove('hidden');
+// 	loading.classList.add('hidden');
+// }
 
 
-//Подписка на кнопку Удалить комменатрий
-function initDeleteButtonEventListeners() {
+// // Подписка на события кнопки Редактировать
+// function initEditButtonEventListeners() {
+// 	const EditButtons = listComments.querySelectorAll('.edit-button');
 
-	const deleteButtons = document.querySelectorAll('.delete-button');
+// 	for (const EditButton of EditButtons) {
+// 		EditButton.addEventListener('click', (event) => {
+// 			event.stopPropagation();
+// 			const target = event.target;
+// 			renderEditComment(target)
+// 		})
+// 	}
+// }
 
-	for (const deleteButton of deleteButtons) {
-		deleteButton.addEventListener('click', (event) => {
-			event.stopPropagation();
-			const index = deleteButton.dataset.index;
-			comments.splice(index, 1);
-			renderComments();
-		})
-	}
-}
+// //Подписка на события клика по кнопке Лайк
+// function initLikeButtonEventListeners() {
 
-//Подписка на событие ответить на комментарий
-function initAnswerCommentEventListener() {
-	const commentElements = document.querySelectorAll('.comment');
+// 	const likeButtons = listComments.querySelectorAll('.like-button');
 
-	for (const commentElement of commentElements) {
-		commentElement.addEventListener('click', (event) => {
-			const index = commentElement.dataset.comment;
+// 	for (const likeButton of likeButtons) {
+// 		likeButton.addEventListener('click', (event) => {
+// 			event.stopPropagation();
+// 			const target = event.target;
+// 			const commentParrent = target.closest('.comment');
+// 			const commentId = commentParrent.dataset.comment;
+// 			likeButton.classList.add('-loading-like');
 
-			quoteBlock = `/** ${comments[index].name}:\n${comments[index].text} **/`;
-
-			inputComment.value = `${quoteBlock}\n\n`;
-
-			inputComment.focus();
-		})
-	}
-}
+// 			delay(2000).then(() => {
+// 				comments[commentId].likes = comments[commentId].isLiked ? comments[commentId].likes -= 1 : comments[commentId].likes += 1;
+// 				comments[commentId].isLiked = !comments[commentId].isLiked;
+// 				renderComments()
+// 			});
+// 		})
+// 	}
+// }
 
 
-// Функция для имитации запросов в API
-function delay(interval = 300) {
-	return new Promise((resolve) => {
-		setTimeout(() => {
-			resolve();
-		}, interval);
-	});
-}
+// //Подписка на кнопку Удалить комменатрий
+// function initDeleteButtonEventListeners() {
+
+// 	const deleteButtons = document.querySelectorAll('.delete-button');
+
+// 	for (const deleteButton of deleteButtons) {
+// 		deleteButton.addEventListener('click', (event) => {
+// 			event.stopPropagation();
+// 			const index = deleteButton.dataset.index;
+// 			comments.splice(index, 1);
+// 			renderComments();
+// 		})
+// 	}
+// }
+
+// //Подписка на событие ответить на комментарий
+// function initAnswerCommentEventListener() {
+// 	const commentElements = document.querySelectorAll('.comment');
+
+// 	for (const commentElement of commentElements) {
+// 		commentElement.addEventListener('click', (event) => {
+// 			const index = commentElement.dataset.comment;
+
+// 			quoteBlock = `/** ${comments[index].name}:\n${comments[index].text} **/`;
+
+// 			inputComment.value = `${quoteBlock}\n\n`;
+
+// 			inputComment.focus();
+// 		})
+// 	}
+// }
+
+
+// // Функция для имитации запросов в API
+// function delay(interval = 300) {
+// 	return new Promise((resolve) => {
+// 		setTimeout(() => {
+// 			resolve();
+// 		}, interval);
+// 	});
+// }
 
 // renderLoading('Подождите, пожалуйста, комментарии загружаются...');
 listComments.textContent = 'Подождите, пожалуйста, комментарии загружаются...';
 initialState();
 
 
-fetchGet()
-	.then(responseData => {
-		const appcomments = responseData.comments.map((comment) => {
-			return {
-				name: comment.author.name,
-				date: renderDate(comment.date),
-				text: comment.text,
-				likes: comment.likes,
-				isLiked: false,
-				forceError: true
-			}
-		})
-		comments = appcomments;
-		renderComments();
-		removerLoading();
-	})
+fetchGet();
+	// .then(responseData => {
+	// 	const appcomments = responseData.comments.map((comment) => {
+	// 		return {
+	// 			name: comment.author.name,
+	// 			date: renderDate(comment.date),
+	// 			text: comment.text,
+	// 			likes: comment.likes,
+	// 			isLiked: false,
+	// 			forceError: true
+	// 		}
+	// 	})
+	// 	comments = appcomments;
+	// 	renderComments();
+	// 	removerLoading();
+	// })
 
 //Подписка на событие клика по кнопке "Написать"
 formButton.addEventListener('click', handlerAddComment);
