+++ variables.js
@@ -1,0 +1,10 @@
+export const form = document.querySelector('.add-form');
+export const listComments = document.querySelector('.comments');
+export const loading = document.querySelector('.load');
+export const inputComment = form.querySelector('.add-form-text');
+export const inputName = form.querySelector('.add-form-name');
+export const formButton = form.querySelector('.add-form-button');
+// export let inputNameValue = '';
+// export let inputCommentValue = '';
+
+
